+++ Documentation/DemoApps/GRDBCombineDemo/GRDBCombineDemo/AppDatabase.swift
@@ -1,4 +1,5 @@
 import Combine
+import Foundation
 import GRDB
 
 /// AppDatabase lets the application access the database.
@@ -95,14 +96,13 @@ extension AppDatabase {
     /// Refresh all players (by performing some random changes, for demo purpose).
     func refreshPlayers() throws {
         try dbWriter.write { db in
-            if try Player.fetchCount(db) == 0 {
+            if try Player.all().isEmpty(db) {
                 // When database is empty, insert new random players
                 try createRandomPlayers(db)
             } else {
                 // Insert a player
                 if Bool.random() {
-                    var player = Player.newRandom()
-                    try player.insert(db)
+                    _ = try Player.makeRandom().inserted(db) // insert but ignore inserted id
                 }
                 
                 // Delete a random player
@@ -123,7 +123,7 @@ extension AppDatabase {
     /// Create random players if the database is empty.
     func createRandomPlayersIfEmpty() throws {
         try dbWriter.write { db in
-            if try Player.fetchCount(db) == 0 {
+            if try Player.all().isEmpty(db) {
                 try createRandomPlayers(db)
             }
         }
@@ -142,8 +142,7 @@ extension AppDatabase {
     func createPlayersForUITests() throws {
         try dbWriter.write { db in
             try AppDatabase.uiTestPlayers.forEach { player in
-                var mutablePlayer = player
-                try mutablePlayer.save(db)
+                _ = try player.inserted(db) // insert but ignore inserted id
             }
         }
     }
@@ -151,14 +150,17 @@ extension AppDatabase {
     /// Support for `createRandomPlayersIfEmpty()` and `refreshPlayers()`.
     private func createRandomPlayers(_ db: Database) throws {
         for _ in 0..<8 {
-            var player = Player.newRandom()
-            try player.insert(db)
+            _ = try Player.makeRandom().inserted(db) // insert but ignore inserted id
         }
     }
 }
 
 // MARK: - Database Access: Reads
 
+// This demo app does not provide any specific reading method, and instead
+// gives an unrestricted read-only access to the rest of the application.
+// In your app, you are free to choose another path, and define focused
+// reading methods.
 extension AppDatabase {
     /// Provides a read-only access to the database
     var databaseReader: DatabaseReader {