+++ Documentation/DemoApps/GRDBCombineDemo/GRDBCombineDemo/Views/PlayerCreationView.swift
@@ -2,37 +2,40 @@ import SwiftUI
 
 /// The view that creates a new player.
 struct PlayerCreationView: View {
-    /// Executed when user cancels or saves the new user.
-    let dismissAction: () -> Void
-    
+    /// Write access to the database
     @Environment(\.appDatabase) private var appDatabase
-    @State private var name = ""
-    @State private var score = ""
+    @Environment(\.dismiss) private var dismiss
+    @State private var form = PlayerForm(name: "", score: "")
     @State private var errorAlertIsPresented = false
     @State private var errorAlertTitle = ""
     
     var body: some View {
         NavigationView {
-            PlayerFormView(name: $name, score: $score)
+            PlayerFormView(form: $form)
                 .alert(
                     isPresented: $errorAlertIsPresented,
                     content: { Alert(title: Text(errorAlertTitle)) })
                 .navigationBarTitle("New Player")
                 .navigationBarItems(
-                    leading: Button(
-                        action: dismissAction,
-                        label: { Text("Cancel") }),
-                    trailing: Button(
-                        action: save,
-                        label: { Text("Save") }))
+                    leading: Button(role: .cancel) {
+                        dismiss()
+                    } label: {
+                        Text("Cancel")
+                    },
+                    trailing: Button {
+                        save()
+                    } label: {
+                        Text("Save")
+                    })
         }
     }
     
     private func save() {
         do {
-            var player = Player(id: nil, name: name, score: Int(score) ?? 0)
-            try appDatabase?.savePlayer(&player)
-            dismissAction()
+            var player = Player(id: nil, name: "", score: 0)
+            form.apply(to: &player)
+            try appDatabase.savePlayer(&player)
+            dismiss()
         } catch {
             errorAlertTitle = (error as? LocalizedError)?.errorDescription ?? "An error occurred"
             errorAlertIsPresented = true
@@ -42,7 +45,6 @@ struct PlayerCreationView: View {
 
 struct PlayerCreationSheet_Previews: PreviewProvider {
     static var previews: some View {
-        PlayerCreationView(dismissAction: { })
-            .environment(\.appDatabase, .empty())
+        PlayerCreationView()
     }
 }