+++ Documentation/DemoApps/GRDBCombineDemo/GRDBCombineDemo/Views/PlayerEditionView.swift
@@ -2,18 +2,28 @@ import SwiftUI
 
 /// The view that edits an existing player.
 struct PlayerEditionView: View {
+    /// Write access to the database
     @Environment(\.appDatabase) private var appDatabase
-    @State var player: Player
+    @Environment(\.isPresented) private var isPresented
+    private let player: Player
+    @State private var form: PlayerForm
+    
+    init(player: Player) {
+        self.player = player
+        self.form = PlayerForm(player)
+    }
     
     var body: some View {
-        PlayerFormView(
-            name: $player.name,
-            score: Binding(
-                get: { "\(player.score)" },
-                set: { player.score = Int($0) ?? 0 }))
-            .onDisappear {
-                // save and ignore error
-                try? appDatabase?.savePlayer(&player)
+        PlayerFormView(form: $form)
+            .onChange(of: isPresented) { isPresented in
+                // Save when back button is pressed
+                if !isPresented {
+                    var savedPlayer = player
+                    form.apply(to: &savedPlayer)
+                    // Ignore error because I don't know how to cancel the
+                    // back button and present the error
+                    try? appDatabase.savePlayer(&savedPlayer)
+                }
             }
     }
 }
@@ -21,7 +31,7 @@ struct PlayerEditionView: View {
 struct PlayerEditionView_Previews: PreviewProvider {
     static var previews: some View {
         NavigationView {
-            PlayerEditionView(player: Player.newRandom())
+            PlayerEditionView(player: Player.makeRandom())
                 .navigationBarTitle("Player Edition")
         }
     }